backend/requirements.txt
fastapi
uvicorn
sentence-transformers
pdfplumber
python-docx
numpy
scikit-learn
python-multipart
